requirements.txt
modal>=0.56.4094
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
pdf2image>=1.16.3 